+++ src/assert.js
@@ -1,7 +1,7 @@
 
 /*@__NO_SIDE_EFFECTS__*/
 export default function assert(val) {
-	//#IFDEV
+	//#IFDEBUG
 	if (!val) {
 		//debugger;
 		throw new Error('Assertion failed: ' + val);